requirements.txt
attrs==21.4.0
click==8.1.2
dnspython==2.2.1
Flask==2.1.1
Flask-Cors==3.0.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
pymongo==4.1.1
pyrsistent==0.18.1
python-dotenv==0.20.0
six==1.16.0
structlog==21.5.0
Werkzeug==2.1.1
zipp==3.8.0
